wu-mqtt-bridge/app/requirements.txt
aiohttp>=3.9,<4
aiomqtt>=2.0,<3
cryptography>=42.0,<47
dnspython>=2.6,<3
